Restructured code to be clearer
