+++ app/IMessage.ts
