+++ app/IMessageQueue.ts
