+++ app/QueuedMessage.ts
