+++ app/hashtable.ts
